voice-server/pyproject.toml
[project]
name = "noctem-voice-server"
version = "0.1.0"
description = "Noctem Voice Server — Twilio + OpenAI Realtime bridge"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]
